code DRY for skeleton + empty state
