+++ centreon/packages/ui/src/Listing/EmptyResult/EmptyResult.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import { EmptyRow } from '../Row/EmptyRow';
+
+interface EmptyResultProps {
+  label: string;
+}
+
+const EmptyResult = ({ label }: EmptyResultProps): JSX.Element => (
+  <EmptyRow>{label}</EmptyRow>
+);
+
+export { EmptyResult };
